requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
